# [Improvement](aggregate) optimization for AggregationMethodKeysFixed::insert_keys_into_columns