## Proposed changes
```sql
select count(1) from (
select ss_customer_sk customer_sk
      ,ss_item_sk item_sk
from store_sales,date_dim
where ss_sold_date_sk = d_date_sk
  and d_month_seq between 1199 and 1199 + 11 and ss_sold_date_sk IS NOT NULL
group by ss_customer_sk
        ,ss_item_sk
) a;
```
```
uint256 original
12.27 sec
                              -  HashTableInputCount:  55.286154M  (55286154)
                              -  HashTableIterateTime:  347.592ms
                              -  HashTableSize:  54.116764M  (54116764)
                              -  InsertKeysToColumnTime:  1s79ms
                              -  MaxRowSizeInBytes:  0
                              -  MemoryUsage:  
                                  -  HashTable:  2.50  GB
                                  -  SerializeKeyArena:  1.63  GB
                              -  MergeTime:  0ns
                              -  PeakMemoryUsage:  4.13  GB

stringref
17.81 sec

                              -  HashTableInputCount:  55.285529M  (55285529)
                              -  HashTableIterateTime:  222.769ms
                              -  HashTableSize:  54.116764M  (54116764)
                              -  InsertKeysToColumnTime:  340.796ms
                              -  MaxRowSizeInBytes:  17
                              -  MemoryUsage:  
                                  -  HashTable:  1.50  GB
                                  -  SerializeKeyArena:  1.76  GB
                              -  MergeTime:  0ns
                              -  PeakMemoryUsage:  3.26  GB

uint256 opt
11.20 sec

                              -  HashTableInputCount:  55.285554M  (55285554)
                              -  HashTableIterateTime:  278.187ms
                              -  HashTableSize:  54.116764M  (54116764)
                              -  InsertKeysToColumnTime:  467.894ms
                              -  MaxRowSizeInBytes:  0
                              -  MemoryUsage:  
                                  -  HashTable:  2.50  GB
                                  -  SerializeKeyArena:  1.63  GB
                              -  MergeTime:  0ns
                              -  PeakMemoryUsage:  4.13  GB

```

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...