@hello-stephen 2023-07-26
(From new machine)TeamCity pipeline, clickbench performance test result:
 the sum of best hot time: 45.55 seconds
 stream load tsv:          508 seconds loaded 74807831229 Bytes, about 140 MB/s
 stream load json:         20 seconds loaded 2358488459 Bytes, about 112 MB/s
 stream load orc:          64 seconds loaded 1101869774 Bytes, about 16 MB/s
 stream load parquet:          32 seconds loaded 861443392 Bytes, about 25 MB/s
 insert into select:          29.9 seconds inserted 10000000 Rows, about 334K ops/s
 storage size: 17156773957 Bytes